+++ src/vmm/src/device_manager/pci_mngr.rs
@@ -672,7 +672,7 @@ mod tests {
     use crate::vmm_config::memory_hotplug::MemoryHotplugConfig;
     use crate::vmm_config::net::NetworkInterfaceConfig;
     use crate::vmm_config::pmem::PmemConfig;
-    use crate::vmm_config::vsock::VsockDeviceConfig;
+    use crate::vmm_config::vsock::{VsockDeviceConfig, VsockType};
 
     #[test]
     fn test_device_manager_persistence() {
@@ -731,6 +731,7 @@ mod tests {
                 vsock_id: Some(vsock_dev_id.to_string()),
                 guest_cid: 3,
                 uds_path: tmp_sock_file.as_path().to_str().unwrap().to_string(),
+                vsock_type: VsockType::Stream,
             };
             insert_vsock_device(&mut vmm, &mut cmdline, &mut event_manager, vsock_config);
             // Add an entropy device.
@@ -840,7 +841,8 @@ mod tests {
   ],
   "vsock": {{
     "guest_cid": 3,
-    "uds_path": "{}"
+    "uds_path": "{}",
+    "vsock_type": "stream"
   }},
   "entropy": {{
     "rate_limiter": null