+++ src/vmm/src/devices/virtio/vsock/unix/muxer_killq.rs
@@ -27,8 +27,10 @@
 use std::collections::{HashMap, VecDeque};
 use std::time::Instant;
 
+use super::defs;
 use super::muxer::ConnMapKey;
-use super::{MuxerConnection, defs};
+use crate::devices::virtio::vsock::csm::VsockConnection;
+use crate::devices::virtio::vsock::unix::ConnBackend;
 
 /// A kill queue item, holding the connection key and the scheduled time for termination.
 #[derive(Debug, Clone, Copy)]
@@ -66,7 +68,7 @@ impl MuxerKillQ {
     /// set to expire at some point in the future.
     /// Note: if more than `Self::SIZE` connections are found, the queue will be created in an
     ///       out-of-sync state, and will be discarded after it is emptied.
-    pub fn from_conn_map(conn_map: &HashMap<ConnMapKey, MuxerConnection>) -> Self {
+    pub fn from_conn_map(conn_map: &HashMap<ConnMapKey, VsockConnection<ConnBackend>>) -> Self {
         let mut q_buf: Vec<MuxerKillQItem> = Vec::with_capacity(Self::SIZE);
         let mut synced = true;
         for (key, conn) in conn_map.iter() {